requirements.txt
pandas
requests
rapidfuzz
streamlit
python-dotenv
plotly